Stable & Wait Free
Relatively large refactor of locking strategy:

1. sessions only modify segments that they allocated
2. sessions sync their dirty segments on transaction commit
      - by sync, I mean mprotect and mark read only
3. compactor and cache promotion only operate on mprotect read-only data

End result is that there is no longer a need to synchronize the
compactor and session threads with respect to modifying in place.
Furthermore, there is no longer a need to synchronize all threads every
time we want to sync, because syncs only impact session/thread-local
segments.

By adopting pro-active mprotect read-only on committed data the database
is protected against bugs in application code writing to database
memory, which is the most likely source of corruption. The consequence
of this is a small increase in COW use, but in reality all committed
revsions of the database have a ref count of 2+ so their modifications
were already COW.